+++ source/core.py
@@ -1,6 +1,6 @@
 # A part of NonVisual Desktop Access (NVDA)
 # Copyright (C) 2006-2026 NV Access Limited, Aleksey Sadovoy, Christopher Toth, Joseph Lee, Peter Vágner,
-# Derek Riemer, Babbage B.V., Zahari Yurukov, Łukasz Golonka, Cyrille Bougot, Julien Cochuyt
+# Derek Riemer, Babbage B.V., Zahari Yurukov, Łukasz Golonka, Cyrille Bougot, Julien Cochuyt, Wang Chong
 # This file is covered by the GNU General Public License.
 # See the file COPYING for more details.
 
@@ -927,6 +927,13 @@ def main():
 	import _remoteClient
 
 	_remoteClient.initialize()
+
+	from textUtils import wordSeg
+
+	try:
+		wordSeg.initialize()
+	except Exception:
+		log.error("Error initializing word segmentation module", exc_info=True)
 
 	if globalVars.appArgs.install or globalVars.appArgs.installSilent:
 		import gui.installerGui